12.1-> Exercise Pokedex State-> Creating directory `components` and moving `Pokedex` and `Pokemon`
