+++ front-end/12.1/exercises/pokedex-state/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import './App.css';
 import pokemons from './data';
-import Pokedex from './Pokedex';
+import Pokedex from './components/Pokedex';
 
 function App() {
   return (
