+++ front-end/12.1/exercises/pokedex-state/src/components/Pokemon.js
